refactor: create config folder and move statusconfig
