+++ src/apps/ScopeChangeRequest/utils/config/index.ts
@@ -1,0 +1,1 @@
+export * from './statusBarConfig';
